+++ src/banks_color.h
@@ -70,8 +70,8 @@ typedef struct color_pal_t {
 
 bool colors_try_windows_enable_virtual_term_for_vt_codes(void);
 
-void bank_render_color(int bank_mem_type, int mode);
+void bank_render_color(bank_item * p_bank, int mode);
 
-bool setoption_custom_bank_colors(char * arg_str);
+bool set_option_custom_bank_colors(char * arg_str);
 
 #endif // _BANKS_PRINT_H